+++ releasenotes.md
@@ -36,6 +36,7 @@
 - Introduce `$vaarg[...]` syntax and deprecate the old `$vaarg(...)`.
 - Similar change to `$vasplat`: `$vasplat` and `$vasplat[1..]`.
 - Add `$member.get(value)` to replace `value.$eval($member.nameof)`
+- Improve the error message when the compilation does not produce any files #1390.
 
 ### Fixes